test(integration): improve query tests and enhance domain modeling
- Added `IncludeInactive` parameter to `GetApartmentsByOwnerQuery` for better query flexibility.
- Updated `TestUserContext` with `IsAdmin` and `IsOwner` domain methods.
- Replaced generic `ImageUrl` and `Avatar` types with domain-specific `ApartmentImageUrl` and `AvatarUrl` for improved type safety.
- Updated assertions in `GetPaymentsByBookingTests` to use domain enums directly instead of casting.

+++ StayHub.Application.IntegrationTests/Payments/GetPaymentByBookingTests.cs
@@ -64,7 +64,7 @@ public class GetPaymentByBookingTests(IntegrationTestWebAppFactory factory) : Ba
         result.IsSuccess.Should().BeTrue();
         result.Value.AmountValue.Should().Be(725.50m);
         result.Value.AmountCurrency.Should().Be("USD");
-        result.Value.Status.Should().Be((int)PaymentStatus.Pending);
+        result.Value.Status.Should().Be(PaymentStatus.Pending);
     }
 
     [Fact]
